fix greedy dirichlet test

+++ packages/muelu/test/unit_tests/Aggregates.cpp
@@ -944,7 +944,7 @@ public:
 
     aggregates->ComputeNodesInAggregate(aggPtr, aggNodes, unaggregated);
     // Test to check that the dirichlet node is aggregated:
-    for( size_t i=0; i < unaggregated.size(); i++){
+    for( int i = 0; i < unaggregated.size(); i++){
       TEST_EQUALITY( unaggregated[i] == 2, false);
     }
